refactor(datafile,writer): address review comments (naming, symmetry, test hygiene)
- Rename the gated `datafile::ext` module to `datafile::datafusion_ext` so its
  purpose (the DataFusion-capable surface) is clear from the name; it keeps the
  `*Ext` framing of the traits it holds. No code imported it (doc-only refs).
- Move the `DataFileReader` trait into `datafile/mod.rs` next to its writer
  counterpart `DataFileWriter` (the other three tier traits already live there),
  making the file/dataset tier definitions symmetric. The impls stay in
  `reader.rs`.
- Reword the kernel-reader "not yet implemented" error: it previously implied
  there is no DataFusion-free read path, but `ParquetTableReader` is one. It now
  points to `ParquetTableReader` for plain tables and the `datafusion` feature
  for full Delta read semantics.
- writer/json.rs: move test-only `use` statements and `extract_partition_values`
  into `mod tests` instead of annotating them `#[cfg(test)]` at module scope;
  drop the redundant "(test-only)" comment. Reword the `write_with_mode` doc to
  describe current behavior rather than narrate the change.
- writer/record_batch.rs: drop the `#[cfg(test)]` `divide_by_partition_values`
  method on the production `RecordBatchWriter`; replace it with a `divide_writer_batch`
  helper inside `mod tests`.
