+++ crates/core/src/datafile/mod.rs
@@ -1,13 +1,13 @@
 //! Data-file read/write abstractions, in two tiers:
 //!
-//! * **File tier** ([`DataFileWriter`], [`reader::DataFileReader`]) — the per-file
+//! * **File tier** ([`DataFileWriter`], [`DataFileReader`]) — the per-file
 //!   seam where parquet `WriterProperties` attach.
 //!   Impl: [`writer::PartitionWriter`].
 //! * **Dataset tier** ([`DeltaDataWriter`], [`DeltaDataReader`]) — composes the
 //!   file tier across a table. Impl: [`writer::DeltaWriter`].
 //!
 //! Both tiers operate on a DataFusion-free stream of [`BatchFuture`]s. The
-//! DataFusion-capable surface lives in the gated [`ext`] module.
+//! DataFusion-capable surface lives in the gated [`datafusion_ext`] module.
 
 use arrow_array::RecordBatch;
 use futures::future::BoxFuture;
@@ -21,7 +21,7 @@ pub mod reader;
 pub mod writer;
 
 #[cfg(feature = "datafusion")]
-pub mod ext;
+pub mod datafusion_ext;
 
 pub use properties::ReaderProperties;
 
@@ -46,7 +46,7 @@ pub fn batches_to_future_stream(batches: Vec<RecordBatch>) -> RecordBatchFutureS
 /// Adapt a fallible `RecordBatch` stream into a [`RecordBatchFutureStream`]:
 /// each item becomes a ready [`BatchFuture`] with its error mapped into
 /// [`DeltaTableError`]. Shared by the parquet file reader ([`reader`]) and the
-/// DataFusion stream adapters ([`ext`]).
+/// DataFusion stream adapters ([`datafusion_ext`]).
 pub fn results_to_future_stream<S, E>(stream: S) -> RecordBatchFutureStream
 where
     S: Stream<Item = Result<RecordBatch, E>> + Send + 'static,
@@ -71,8 +71,19 @@ pub trait DataFileWriter: Send {
     async fn close(self: Box<Self>) -> DeltaResult<Vec<Add>>;
 }
 
+/// File tier: reads a single parquet data file (the per-file decryption seam,
+/// mirroring [`DataFileWriter`]). Impl: [`reader::ParquetFileReader`].
+#[async_trait::async_trait]
+pub trait DataFileReader: Send + Sync {
+    /// Read the parquet data file at `path` into a stream of record batches.
+    async fn read_file(
+        &self,
+        path: object_store::path::Path,
+    ) -> DeltaResult<RecordBatchFutureStream>;
+}
+
 /// Options for a basic (DataFusion-free) read. Richer predicate/projection
-/// pushdown is the DataFusion extension's job ([`ext::DeltaDataReaderExt`]).
+/// pushdown is the DataFusion extension's job ([`datafusion_ext::DeltaDataReaderExt`]).
 #[derive(Debug, Default, Clone)]
 pub struct ReadOptions {
     /// Project to this subset of (logical) column names. `None` reads all columns.
@@ -107,7 +118,7 @@ pub trait DeltaDataWriter: Send {
 }
 
 /// Dataset tier: a DataFusion-free reader that composes the file tier
-/// ([`reader::DataFileReader`]) across a table's data files, applying deletion
+/// ([`DataFileReader`]) across a table's data files, applying deletion
 /// vectors, partition values, and column-mapping transforms.
 #[async_trait::async_trait]
 pub trait DeltaDataReader: Send + Sync {
